konfusion-test-utils: add way to run Zot registry
Zot [1] is an OCI-compliant, lightwheight registry implementation.

We will use it to stand up our own registry instance for integration
tests so that we're not dependent on having push access to some external
registry service.

Add a 'konfusion-test-utils run-zot-registry' command so that devs can
trivially run Zot locally.

Structure the code in such a way that integration tests can easily start
the Zot registry if needed.

Generate a CA certificate and use it to generate a server cert for Zot.
This allows us to enable https and avoid using http for tests
(if we wanted to use http, we would then have to implement insecure
options just for tests).

[1]: https://zotregistry.dev/

Signed-off-by: Adam Cmiel <acmiel@redhat.com>